# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RunSafe API — SMB Operations & Compliance SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sops:
    get:
      operationId: getSops
      tags: [sops]
      summary: List all SOPs in the workspace
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of SOPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sop"
    post:
      operationId: createSop
      tags: [sops]
      summary: Create a new SOP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSopBody"
      responses:
        "201":
          description: Created SOP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sop"
  /sops/{sopId}:
    get:
      operationId: getSop
      tags: [sops]
      summary: Get a single SOP with steps
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SOP with steps
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SopDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSop
      tags: [sops]
      summary: Update a SOP
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSopBody"
      responses:
        "200":
          description: Updated SOP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sop"
    delete:
      operationId: deleteSop
      tags: [sops]
      summary: Delete a SOP
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: SOP deleted
  /sops/{sopId}/steps:
    post:
      operationId: addSopStep
      tags: [sops]
      summary: Add a step to a SOP
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSopStepBody"
      responses:
        "201":
          description: Step created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SopStep"
    put:
      operationId: reorderSopSteps
      tags: [sops]
      summary: Reorder steps in a SOP
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderSopStepsBody"
      responses:
        "200":
          description: Steps reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SopStep"
  /sops/{sopId}/steps/{stepId}:
    patch:
      operationId: updateSopStep
      tags: [sops]
      summary: Update a SOP step
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
        - name: stepId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSopStepBody"
      responses:
        "200":
          description: Updated step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SopStep"
    delete:
      operationId: deleteSopStep
      tags: [sops]
      summary: Delete a SOP step
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
        - name: stepId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Step deleted
  /sops/{sopId}/assign:
    post:
      operationId: assignSop
      tags: [sops]
      summary: Assign a SOP to a team member as a task
      parameters:
        - name: sopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignSopBody"
      responses:
        "201":
          description: Task assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssignment"

  # ─── TASK ASSIGNMENTS ────────────────────────────────────────────
  /tasks:
    get:
      operationId: getTaskAssignments
      tags: [tasks]
      summary: List task assignments (filtered by status or assignee)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: assigneeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of task assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskAssignment"
  /tasks/{taskId}:
    patch:
      operationId: updateTaskAssignment
      tags: [tasks]
      summary: Update a task assignment (status, progress)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskAssignmentBody"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAssignment"
  /tasks/{taskId}/steps/{stepId}/complete:
    post:
      operationId: completeTaskStep
      tags: [tasks]
      summary: Mark a step as complete within a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
        - name: stepId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Step marked complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStepCompletion"

  # ─── COMPLIANCE ──────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── WORKSPACE ───────────────────────────────────────────────
...
    Sop:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        stepCount:
          type: integer
        createdByClerkId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - title
        - stepCount
        - createdByClerkId
        - createdAt
        - updatedAt
    SopDetail:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        createdByClerkId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/SopStep"
      required:
        - id
        - workspaceId
        - title
        - createdByClerkId
        - createdAt
        - updatedAt
        - steps
    SopStep:
      type: object
      properties:
        id:
          type: integer
        sopId:
          type: integer
        orderIndex:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sopId
        - orderIndex
        - title
        - createdAt
    CreateSopBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
      required:
        - title
    UpdateSopBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
    CreateSopStepBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        orderIndex:
          type: integer
      required:
        - title
    UpdateSopStepBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        orderIndex:
          type: integer
    ReorderSopStepsBody:
      type: object
      properties:
        stepIds:
          type: array
          items:
            type: integer
      required:
        - stepIds
    AssignSopBody:
      type: object
      properties:
        assigneeId:
          type: integer
        dueDate:
          type: ["string", "null"]
          format: date
        notes:
          type: string
      required:
        - assigneeId

    # ─── TASK ASSIGNMENTS ────────────────────────────────────────
    TaskAssignment:
      type: object
      properties:
        id:
          type: integer
        sopId:
          type: integer
        sopTitle:
          type: string
        workspaceId:
          type: integer
        assigneeId:
          type: integer
        assigneeName:
          type: string
        assigneeEmail:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed]
        dueDate:
          type: ["string", "null"]
          format: date
        completedAt:
          type: ["string", "null"]
          format: date-time
        stepsTotal:
          type: integer
        stepsCompleted:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sopId
        - sopTitle
        - workspaceId
        - assigneeId
        - assigneeName
        - assigneeEmail
        - status
        - stepsTotal
        - stepsCompleted
        - createdAt
    UpdateTaskAssignmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed]
        notes:
          type: string
        dueDate:
          type: string
          format: date
    TaskStepCompletion:
      type: object
      properties:
        id:
          type: integer
        taskAssignmentId:
          type: integer
        sopStepId:
          type: integer
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - taskAssignmentId
        - sopStepId
        - completedAt

    # ─── COMPLIANCE ──────────────────────────────────────────────
...