# lib/api-zod/src/generated/types/sopDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunSafe API — SMB Operations & Compliance SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { SopStep } from "./sopStep";

export interface SopDetail {
  id: number;
  workspaceId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  createdByClerkId: string;
  createdAt: Date;
  updatedAt: Date;
  steps: SopStep[];
}