# lib/api-zod/src/generated/types/taskAssignment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunSafe API — SMB Operations & Compliance SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { TaskAssignmentStatus } from "./taskAssignmentStatus";

export interface TaskAssignment {
  id: number;
  sopId: number;
  sopTitle: string;
  workspaceId: number;
  assigneeId: number;
  assigneeName: string;
  assigneeEmail: string;
  status: TaskAssignmentStatus;
  /** @nullable */
  dueDate?: Date | null;
  /** @nullable */
  completedAt?: Date | null;
  stepsTotal: number;
  stepsCompleted: number;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}