# lib/api-zod/src/generated/types/taskStepCompletion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunSafe API — SMB Operations & Compliance SaaS
 * OpenAPI spec version: 0.1.0
 */

export interface TaskStepCompletion {
  id: number;
  taskAssignmentId: number;
  sopStepId: number;
  completedAt: Date;
}