# lib/api-zod/src/generated/types/updateSopBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RunSafe API — SMB Operations & Compliance SaaS
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSopBody {
  title?: string;
  description?: string;
  category?: string;
}